Add command to compile weekly max points
- add a command to compile weekly max points per team
- to calculate the roster's max points for a given week we can just use
a greedy algorithm where we sort the players by their score in descending
order and iterate through the sorted list and place the player in the
available roster position. this guarantees that the roster position is
filled with the highest possible scoring player. if a position is filled,
we check if the player is eligible for the FLEX spot and fill as able
- in special circumstances, a player can be eligible for multiple positions
(as is the case for week 11 roster id 9)
- we can handle this situation by modifying our greedy algorithm to sort
players on 2 criteria: 1. number of eligible positions, ascending 2. number
of points scored descending. this puts multi-positional players at the end
of our sorted array so that we can greedy fill our roster and then we can
iterate our multi-positional players to calculate the biggest point differential
between eligible roster spots and replace the given player based on the
point differences.