Update compile players data command
- sleeper api does not provide historical player data so that we can track
changes so we need to update the command so that we keep old records of players
in order to be able to use old player data since we are presenting historical
data that partially rely on things like a player's eligible positions at the
time
- we cannot, however, simply create new records each time we pull players
because that means even if we do this weekly, we are pulling and adding
over 10k documents to mongo...
- my current idea is to pull player data from sleeper, pull player data from
mongo group it by id and filter to the max year and max week record we have
for the player. theory being that if we need a player's data we just find
the most recent record - highest year, highest week - even if it is not
the current year and/or week the most recent record means that was the
last change detected on the player and is up to date as of the time of
querying the documents.
- since player data is going to be largely static for the fields that we
(currently) care about, this should greatly limit the amount of documents
we add to our mongo collection each week when we run the data compilation.