Add command to compile projections data
- add a command to compile players projected rank data
- this data didn't seem to be readily available or easily findable on sleeper
the best source i found was on espn that seemed to generally line up with
my expectations given point system for the league
- espn doesn't really have an easily accessible api, though they technically
do have a public api. instead i scraped their projections pages to get
players' projected ranks
- this command reads in the file saved from the scraped data and tries to
map players by espn id (this id is generally available in the player data
from sleeper's api). if an espn id match is not found, a match by name is
attempted. finally if a name match doesn't exist and the espn player name
is greater than 2 words (most of the time due to a suffix like Jr., II, etc.)
then we try to match by just the first 2 parts of the name since it also
seemed like a lot of the time sleeper does not include suffixes in names
- the command will update player records with their espn id if needed
- finally the command will write player projected rankings to mongodb