Update compile player projections matching logic
- update the logic when we match multiple players
- our query for when we match by source_id OR full_name was intended to
use full_name as a fall back if we didn't find a match by source_id but
that's not quite how it works. we would have to do multiple queries to
truly only use full_name if source_id didn't return a match.
- since our match will return anyone who matches source_id OR full_name
we need to check if any of our matches match the source_id before we consider
the matches to simply be name duplicates
- this will fix big players with source_id such as Josh Allen, Lamar Jackson, etc.
to properly map to their projection instead of showing as 'no projection'.
no projection data should only really ever happen on fringe borderline nfl players