Update the Points Scored column
- In the league summary table update the points scored column to show
a gauge chart to visualize a team's actual total poitns scored compared
to their potential minimum and potential maximum scores
- This implementation runs into a visualization issue where the scales
for each row are independent of each other but because the charts are the
same size and the same gradient colors they seem like they represent identical
ranges. this can be misleading. we'll need to come up with a better way
to visualize this probably
- remove the potential max column since that information is now conveyed
in the new points scored column